migrate to overload approach
