+++ patches/net/minecraft/world/item/crafting/RecipeManager.java.patch
@@ -8,12 +8,18 @@
  
      public RecipeManager(HolderLookup.Provider registries) {
          this.registries = registries;
-@@ -69,7 +_,7 @@
+@@ -68,8 +_,13 @@
+ 
      protected RecipeMap prepare(ResourceManager manager, ProfilerFiller profiler) {
          SortedMap<Identifier, Recipe<?>> recipes = new TreeMap<>();
++        // Neo: add condition context and post recipe JSON modification event
++        var conditionalOps = new net.neoforged.neoforge.common.conditions.ConditionalOps<>(this.registries.createSerializationContext(JsonOps.INSTANCE), getContext());
          SimpleJsonResourceReloadListener.scanDirectory(
 -            manager, RECIPE_LISTER, this.registries.createSerializationContext(JsonOps.INSTANCE), Recipe.CODEC, recipes
-+            manager, RECIPE_LISTER, new net.neoforged.neoforge.common.conditions.ConditionalOps<>(this.registries.createSerializationContext(JsonOps.INSTANCE), getContext()), Recipe.CODEC, recipes // Neo: add condition context
++            manager, RECIPE_LISTER, conditionalOps, Recipe.CODEC, recipes, recipeJsons -> {
++                var event = new net.neoforged.neoforge.data.event.ModifyRecipeJsonEvent(conditionalOps.lookupProvider, recipeJsons);
++                net.neoforged.fml.ModLoader.postEventWrapContainerInModOrder(event);
++            }
          );
          List<RecipeHolder<?>> recipeHolders = new ArrayList<>(recipes.size());
          recipes.forEach((id, recipe) -> {
